# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZTAP REST API
  version: v1
  description: |
    Minimal REST API for ZTAP.

    Auth uses bearer tokens from `POST /v1/auth/login`.
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /v1/policies:
    get:
      tags: [policies]
      summary: List policies
      security:
        - bearerAuth: []
      parameters:
        - name: tenant
          in: query
          schema:
            type: string
          description: Optional tenant filter
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliciesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
  /v1/policies/{tenant}/{name}:
    get:
      tags: [policies]
      summary: Get a policy
      security:
        - bearerAuth: []
      parameters:
        - name: tenant
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    put:
      tags: [policies]
      summary: Create or update a policy
      security:
        - bearerAuth: []
      parameters:
        - name: tenant
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyUpsertRequest"
      responses:
        "200":
          description: Policy updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyUpsertResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    delete:
      tags: [policies]
      summary: Delete a policy
      security:
        - bearerAuth: []
      parameters:
        - name: tenant
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyDeleteRequest"
      responses:
        "200":
          description: Policy deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyDeleteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
  /v1/cluster/nodes:
    get:
      tags: [cluster]
      summary: List cluster nodes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nodes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterNodesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          description: Cluster not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
    post:
      tags: [cluster]
      summary: Register a cluster node
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClusterNodeRequest"
      responses:
        "201":
          description: Node registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterNode"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "501":
          description: Cluster not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PolicyInfo:
      type: object
      properties:
        tenant:
          type: string
        name:
          type: string
        version:
          type: integer
          format: int64
        source:
          type: string
        updated_at:
          type: string
          format: date-time
        policy_yaml:
          type: string
      required: [tenant, name, version, source, updated_at]
    PoliciesResponse:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: "#/components/schemas/PolicyInfo"
      required: [policies]
    PolicyUpsertRequest:
      type: object
      properties:
        policy_yaml:
          type: string
        reason:
          type: string
        expected_version:
          type: integer
          format: int64
          minimum: 0
      required: [policy_yaml]
    PolicyUpsertResponse:
      type: object
      properties:
        version:
          type: integer
          format: int64
      required: [version]
    PolicyDeleteRequest:
      type: object
      properties:
        reason:
          type: string
    PolicyDeleteResponse:
      type: object
      properties:
        version:
          type: integer
          format: int64
        deleted:
          type: boolean
      required: [version, deleted]
...
    ClusterNode:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        role:
          type: string
        state:
          type: string
        joined_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
      required: [id, address, state, joined_at, last_seen]
    ClusterNodeRequest:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        role:
          type: string
        state:
          type: string
        joined_at:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties:
            type: string
      required: [id, address]
...
    ClusterNodesResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/ClusterNode"
      required: [nodes]
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    MethodNotAllowed:
      description: Method not allowed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"